+++ gulpfile.js
@@ -58,7 +58,9 @@ const doCompile = function (buildNls) {
 		.pipe(sourcemaps.init())
 		.pipe(tsProject()).js
 		.pipe(buildNls ? nls.rewriteLocalizeCalls() : es.through())
-		.pipe(buildNls ? nls.createAdditionalLanguageFiles(languages, 'i18n', 'out') : es.through());
+		.pipe(buildNls ? nls.createAdditionalLanguageFiles(languages, 'i18n', 'out') : es.through())
+		.pipe(buildNls ? nls.bundleMetaDataFiles('vscode-textile-preview', 'out') : es.through())
+		.pipe(buildNls ? nls.bundleLanguageFiles() : es.through());
 
 	if (inlineMap && inlineSource) {
 		r = r.pipe(sourcemaps.write());