+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Udate translations from vscode-loc repo
 
+### Fixed
+- [INTERNAL] vscode-nls : use 1 file bundle
+
 
 ## [0.2.5] - 2019-09-08
 ### Added